CSSTUDIO-1362 Update log search UI
