+++ app/logbook/ui/src/main/resources/org/phoebus/logbook/ui/AdvancedSearchView.fxml
@@ -1,0 +1,102 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+
+<!--
+  ~ Copyright (C) 2020 European Spallation Source ERIC.
+  ~
+  ~  This program is free software; you can redistribute it and/or
+  ~  modify it under the terms of the GNU General Public License
+  ~  as published by the Free Software Foundation; either version 2
+  ~  of the License, or (at your option) any later version.
+  ~
+  ~  This program is distributed in the hope that it will be useful,
+  ~  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  ~  GNU General Public License for more details.
+  ~
+  ~  You should have received a copy of the GNU General Public License
+  ~  along with this program; if not, write to the Free Software
+  ~  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+  -->
+
+<?import javafx.geometry.Insets?>
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="org.phoebus.logbook.ui.AdvancedSearchViewController"
+            fx:id="advancedSearchPane">
+
+        <children>
+            <GridPane layoutX="18.0" layoutY="95.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+                <columnConstraints>
+                    <ColumnConstraints hgrow="NEVER" minWidth="10.0" prefWidth="100.0" />
+                    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                </columnConstraints>
+                <rowConstraints>
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
+                    <RowConstraints />
+                    <RowConstraints />
+                </rowConstraints>
+                <children>
+                    <Label text="Title:" GridPane.rowIndex="0"/>
+                    <TextField GridPane.columnSpan="2" GridPane.rowIndex="1" />
+                    <Label text="Text:" GridPane.rowIndex="2"/>
+                    <TextField fx:id="searchText" onAction="#setSearchText" GridPane.columnSpan="2" GridPane.rowIndex="3"/>
+                    <Label text="Level:" GridPane.rowIndex="4" />
+                    <TextField GridPane.columnSpan="2" GridPane.rowIndex="5" />
+                    <Label text="Logbooks:" GridPane.rowIndex="6"/>
+                    <TextField fx:id="searchLogbooks" GridPane.columnSpan="2" GridPane.rowIndex="7"/>
+                    <Label text="Tags:" GridPane.rowIndex="8"/>
+                    <TextField fx:id="searchTags" GridPane.columnSpan="2" GridPane.rowIndex="9"/>
+                    <Label text="Author:" GridPane.rowIndex="10" />
+                    <TextField fx:id="searchAuthor" GridPane.columnSpan="2" GridPane.rowIndex="11"/>
+                    <Label text="Time:" GridPane.columnSpan="2" GridPane.rowIndex="12">
+                        <GridPane.margin>
+                            <Insets top="5.0" />
+                        </GridPane.margin>
+                    </Label>
+
+                    <GridPane fx:id="timePane" GridPane.columnSpan="2" GridPane.rowIndex="13">
+                        <columnConstraints>
+                            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                        </columnConstraints>
+                        <rowConstraints>
+                            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                        </rowConstraints>
+                        <children>
+                            <Label text="Start Time:" GridPane.rowIndex="0">
+                                <GridPane.margin>
+                                    <Insets />
+                                </GridPane.margin></Label>
+                            <TextField fx:id="startTime" GridPane.columnIndex="1" GridPane.hgrow="ALWAYS" GridPane.rowIndex="0">
+                                <GridPane.margin>
+                                    <Insets />
+                                </GridPane.margin></TextField>
+                            <Label text="End Time:" GridPane.rowIndex="1" />
+                            <TextField fx:id="endTime" GridPane.columnIndex="1" GridPane.hgrow="ALWAYS" GridPane.rowIndex="1">
+                                <GridPane.margin>
+                                    <Insets />
+                                </GridPane.margin></TextField>
+                        </children>
+                    </GridPane>
+                </children>
+            </GridPane>
+        </children>
+        <padding>
+            <Insets left="5.0" right="5.0" />
+        </padding>
+    </AnchorPane>
